move constants to utils
