+++ clmm/cosmology/parent_class.py
@@ -4,7 +4,7 @@ CLMMCosmology abstract class
 # CLMM Cosmology object abstract superclass
 import numpy as np
 from ..utils import validate_argument, compute_for_good_redshifts
-from ..constants import Constants as const
+from ..utils.constants import Constants as const
 
 
 class CLMMCosmology:
