+++ clmm/utils/validation.py
@@ -1,7 +1,7 @@
 """General utility functions that are used in multiple modules"""
 
 import numpy as np
-from ..constants import Constants as const
+from ..utils.constants import Constants as const
 
 
 def arguments_consistency(arguments, names=None, prefix=""):
@@ -226,6 +226,7 @@ def _validate_is_deltasigma_sigma_c(is_deltasigma, sigma_c):
     if not is_deltasigma and sigma_c is not None:
         raise TypeError(f"sigma_c (={sigma_c}) must be None when is_deltasigma=False")
 
+
 def _validate_coordinate_system(loc, coordinate_system, valid_type):
     r"""Validate the coordinate system.
 
@@ -244,6 +245,7 @@ def _validate_coordinate_system(loc, coordinate_system, valid_type):
     validate_argument(loc, coordinate_system, valid_type)
     if loc[coordinate_system] not in ["celestial", "euclidean"]:
         raise ValueError(f"{coordinate_system} must be 'celestial' or 'euclidean'.")
+
 
 class DiffArray:
     """Array where arr1==arr2 is actually all(arr1==arr)"""
